requirements.txt
numpy
pandas
requests
scikit-learn
streamlit
spotipy
sqlite-database
python-dotenv